// music/src/interfaces/pages.interface.ts
export enum EPages {
  HOME = 'Home',
  ARTISTS = 'Artists',
  RELEASES = 'Releases',
  EVENTS = 'Events',
  PODCASTS = 'Podcasts',
  PAGES = 'Pages',
  STORE = 'Store',
  NEWS = 'News',
  PRODUCT = 'Product',
  ARTICLE = 'Article',
  CART = 'Cart',
  PROFILE = 'Profile',
  ABOUT = 'About',
  CONTACTS = 'Contacts',
  PRICING_PLANS = 'Pricing plans',
  PRIVACY_POLICY = 'Privacy policy',
  SIGN_IN = 'Sign in',
  SIGN_UP = 'Sign up',
  FORGOT_PASSWORD = 'Forgot password',
  PAGE_404 = '404 Page'
}